+++ packages/core/core/src/warning-utils.ts
@@ -9,22 +9,24 @@ export enum Codes {
   VERWAR002 = 'VERWAR002',
   VERWAR003 = 'VERWAR003',
   VERWAR004 = 'VERWAR004',
-  VERWAR005 = 'VERWAR005',
   // deprecation warnings
   VERDEP003 = 'VERDEP003',
   VERWAR006 = 'VERWAR006',
+  VERWAR007 = 'VERWAR007',
 }
 
-warningInstance.create(
-  verdaccioWarning,
-  Codes.VERWAR002,
-  `The configuration property "logs" has been deprecated, please rename to "log" for future compatibility`
-);
+/* general warnings */
 
 warningInstance.create(
   verdaccioWarning,
   Codes.VERWAR001,
   `Verdaccio doesn't need superuser privileges. don't run it under root`
+);
+
+warningInstance.create(
+  verdaccioWarning,
+  Codes.VERWAR002,
+  `The configuration property "logs" has been deprecated, please rename to "log" for future compatibility`
 );
 
 warningInstance.create(
@@ -42,21 +44,24 @@ https://verdaccio.org/docs/en/configuration#listen-port`
 );
 
 warningInstance.create(
-  verdaccioWarning,
-  Codes.VERWAR005,
-  'disable enhanced legacy signature is considered a security risk, please reconsider enable it'
+  verdaccioDeprecation,
+  Codes.VERWAR006,
+  'the auth plugin method "add_user" in the auth plugin is deprecated and will be removed in next major release, rename to "adduser"'
 );
+
+warningInstance.create(
+  verdaccioDeprecation,
+  Codes.VERWAR007,
+  `the secret length is too long, it must be 32 characters long, please consider generate a new one 
+  Learn more at https://verdaccio.org/docs/configuration/#.verdaccio-db`
+);
+
+/* deprecation warnings */
 
 warningInstance.create(
   verdaccioDeprecation,
   Codes.VERDEP003,
   'multiple addresses will be deprecated in the next major, only use one'
-);
-
-warningInstance.create(
-  verdaccioDeprecation,
-  Codes.VERWAR006,
-  'the auth plugin method "add_user" in the auth plugin is deprecated and will be removed in next major release, rename to "adduser"'
 );
 
 export function emit(code: string, a?: string, b?: string, c?: string) {